Removed ZooKeeper and Curator dependencies
