--- src/main/java/kafdrop/service/ConsumerMonitor.java
@@ -1,9 +1,0 @@
-package kafdrop.service;
-
-import kafdrop.model.*;
-
-import java.util.*;
-
-public interface ConsumerMonitor {
-  List<ConsumerVO> getConsumers(Collection<TopicVO> topicVos);
-}
